requirements.txt
asgiref==3.5.0
autopep8==1.6.0
black==22.3.0
businesstimedelta==1.0.1
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
convertdate==2.4.0
cryptography==37.0.2
defusedxml==0.7.1
dj-rest-auth==2.2.4
Django==4.0.4
django-allauth==0.50.0
django-cors-headers==3.11.0
django-environ==0.8.1
django-extra-fields==3.0.2
django-seed==0.3.1
django-softdelete==0.9.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
Faker==13.12.0
hijri-converter==2.2.4
holidays==0.14.2
idna==3.3
korean-lunar-calendar==0.2.1
mypy-extensions==0.4.3
mysqlclient==2.1.1
ndg-httpsclient==0.5.1
oauthlib==3.2.0
pathspec==0.9.0
Pillow==9.1.1
platformdirs==2.5.2
# psycopg2==2.9.3
psycopg2-binary==2.9.3
pusher==3.2.0
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
PyMeeus==0.5.11
PyNaCl==1.5.0
pyOpenSSL==22.0.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
toposort==1.7
tzdata==2022.1
urllib3==1.26.9
vscode-ext==1.5.4
